+++ src/main/java/wooteco/subway/exception/badrequest/DuplicateEmailException.java
@@ -1,7 +1,7 @@
 package wooteco.subway.exception.badrequest;
 
-public class DuplicateEmailException extends BadRequest{
-    private static final String MESSAGE = "이미 가입된 이메일입니다.";
+public class DuplicateEmailException extends BadRequest {
+    private static final String MESSAGE = "이미 가입된 이메일입니다";
 
     public DuplicateEmailException() {
         super(MESSAGE);
