+++ src/main/java/wooteco/subway/exception/notfound/MemberNotFoundException.java
@@ -1,7 +1,7 @@
 package wooteco.subway.exception.notfound;
 
 public class MemberNotFoundException extends NotFoundException {
-    private static final String MESSAGE = "존재하지 않는 회원입니다.";
+    private static final String MESSAGE = "존재하지 않는 회원입니다";
 
     public MemberNotFoundException() {
         super(MESSAGE);
